+++ spec/inputs/s3_spec.rb
@@ -428,11 +428,18 @@ describe LogStash::Inputs::S3 do
       include_examples "generated events"
     end
 
-    context 'compressed with gzip extension' do
+    context 'compressed with gzip extension and using default gzip_pattern option' do
       let(:log) { double(:key => 'log.gz', :last_modified => Time.now - 2 * day, :content_length => 5, :storage_class => 'STANDARD') }
       let(:log_file) { File.join(File.dirname(__FILE__), '..', 'fixtures', 'compressed.log.gzip') }
 
       include_examples "generated events"
+    end
+
+    context 'compressed with gzip extension and using custom gzip_pattern option' do
+      let(:config) { super.merge({ "gzip_pattern" => "gee.zip$" }) }
+      let(:log) { double(:key => 'log.gee.zip', :last_modified => Time.now - 2 * day, :content_length => 5, :storage_class => 'STANDARD') }
+      let(:log_file) { File.join(File.dirname(__FILE__), '..', 'fixtures', 'compressed.log.gee.zip') }
+       include_examples "generated events"
     end
 
     context 'plain text' do